Adicionando validações para quando a entidade não existir no banco de dados em CandidatoDa e AvaliadorDao

+++ rhecruta-java/src/main/java/io/github/dac/rhecruta/dao/implementations/AvaliadorDao.java
@@ -43,7 +43,9 @@ public class AvaliadorDao implements AvaliadorDaoInterface {
     @Override
     public void remover(Avaliador avaliador) {
         Avaliador avaliadorToRemove = entityManager.find(Avaliador.class, avaliador.getCpf());
-        entityManager.remove(avaliadorToRemove);
+
+        if (avaliadorToRemove != null)
+            entityManager.remove(avaliadorToRemove);
     }
 
     @Override
